please migrate this class to Azure Functions. with Azure Storage Account as replacement of AWS S3.
In the `run` method, we use the `@BlobTrigger` annotation to indicate that this method should be triggered when a new blob is added to the `input-container` in Azure Blob Storage. The `content` parameter will contain the binary data of the uploaded blob, and the `name` parameter will contain the name of the blob.

We then follow a similar logic as before: infer the image type, resize the image, re-encode it, and upload it to Azure Blob Storage using the Azure Storage SDK.

Note: Replace `"your-storage-connection-string"` with the connection string of your Azure Storage account.

Make sure to configure the input and output container names in the function.json file. The input container should be configured with the path `"input-container/{name}"`, and the output container should be configured with the path `"output-container/{name}"`.

Once you have made these modifications, you can deploy the Azure Function and it should be triggered whenever a new blob is added to the input container. The resized image will be stored in the output container in Azure Blob Storage.

Remember to replace the placeholders with your actual Azure Storage account details and container names.